requirements.tests.txt
# Strictly for tests
coverage
pytest
pytest-asyncio
pytest-cov
pytest-homeassistant-custom-component

homeassistant==2025.3.1
aiohttp_cors